Add decryption gadget
