+++ src/encryption/elgamal.rs
@@ -34,27 +34,39 @@ pub fn decrypt(
     ciphertext_1: &JubJubExtended,
     ciphertext_2: &JubJubExtended,
 ) -> JubJubExtended {
+    // return the plaintext
     ciphertext_2 - ciphertext_1 * secret_key
 }
 
-/// Encrypt in-circuit a plaintext, returning
+/// Encrypt in-circuit a plaintext WitnessPoint, returning
 /// a ciphertext (WitnessPoint, WitnessPoint)
 #[cfg(feature = "zk")]
-pub fn zk_encrypt(
+pub fn encrypt_gadget(
     composer: &mut Composer,
-    public_key: &JubJubAffine,
-    plaintext: &JubJubAffine,
-    r: &JubJubScalar,
+    public_key: WitnessPoint,
+    plaintext: WitnessPoint,
+    r: Witness,
 ) -> Result<(WitnessPoint, WitnessPoint), Error> {
-    // IMPORT INPUTS
-    let public_key = composer.append_point(*public_key);
-    let plaintext = composer.append_point(*plaintext);
-    let r = composer.append_witness(*r);
-
-    // ENCRYPT
     let R = composer.component_mul_point(r, public_key);
     let ciphertext_1 = composer.component_mul_generator(r, GENERATOR)?;
     let ciphertext_2 = composer.component_add_point(plaintext, R);
 
     Ok((ciphertext_1, ciphertext_2))
+}
+
+/// Decrypt in-circuit a ciphertext (WitnessPoint, WitnessPoint),
+/// returning a plaintext WitnessPoint
+#[cfg(feature = "zk")]
+pub fn decrypt_gadget(
+    composer: &mut Composer,
+    secret_key: Witness,
+    ciphertext_1: WitnessPoint,
+    ciphertext_2: WitnessPoint,
+) -> WitnessPoint {
+    let c1_sk = composer.component_mul_point(secret_key, ciphertext_1);
+    let neg_one = composer.append_constant(-JubJubScalar::one());
+    let neg_c1_sk = composer.component_mul_point(neg_one, c1_sk);
+
+    // return plaintext
+    composer.component_add_point(ciphertext_2, neg_c1_sk)
 }
